@@ -7,7 +7,7 @@ name = "mbodied"
 dynamic = ["version"]
 description = 'Embodied AI'
 readme = "README.md"
-requires-python = ">=3.9"
+requires-python = ">=3.10"
 license = "apache-2.0"
 keywords = []
 authors = [{ name = "Mbodi Team", email = "info@mbodi.ai" }]
@@ -81,7 +81,7 @@ dependencies = ["pytest", "pytest-mock", "pytest-asyncio"]
 [tool.hatch.envs.default.env-vars]
 [tool.hatch.envs.conda]
 type = "conda"
-python = "310"                       # Python version 3.10
+python = "311"                       # Python version 3.10
 command = "conda"                    # Options: conda, mamba, micromamba
 conda-forge = false                  # Use conda-forge index
 environment-file = "environment.yml" # Specify environment file
@@ -117,7 +117,7 @@ exclude_lines = ["no cov", "if __name__ == .__main__.:", "if TYPE_CHECKING:"]
 [tool.ruff]
 line-length = 120
 indent-width = 4
-target-version = "py310"
+target-version = "py311"
 
 [tool.ruff.lint]
 extend-unsafe-fixes = ["ALL"]
